Sandbox utility
once in a while for unit testing, you might have a larger module splitup
in many files, and thus you theoretically can unit test each separate
function (which is possible the main usage of unit testing.

Adding  a helper function that reads a number of javascript files,
concatenates them and evaluate them in a sandbox, and returns it's
context.

See tests for example.