Added assert.match to test text for regex
Sometimes it's useful to test if some returned text
matches a specific regex. It may be a dynamic time which
should be a certain structure, but the exact values are not
important to the outcome of the test. Regular Expression
are already widely used and common in JavaScript and can be
very powerful from a testing point of view aswell.